+++ contracts/access/Ownable.sol
@@ -1,38 +1,77 @@
 // SPDX-License-Identifier: MIT
-// OpenZeppelin Contracts (last updated v5.0.0) (access/Ownable.sol)
 
 pragma solidity ^0.8.20;
 
 import {Context} from "../utils/Context.sol";
-import {ERC173} from "./ERC173.sol";
+import {IERC173} from "../interfaces/IERC173.sol";
+import {IERC165, ERC165} from "../utils/introspection/ERC165.sol";
 
 /**
- * @dev Variant of {ERC173} that restricts {transferOwnership} to only
- * receive a non-zero address. The {renounceOwnership} function is provided as
- * a replacement.
+ * @dev Contract module which provides a basic access control mechanism, where
+ * there is an account (an owner) that can be granted exclusive access to
+ * specific functions.
+ *
+ * The initial owner is set to the address provided by the deployer. This can
+ * later be changed with {transferOwnership}.
+ *
+ * This module is used through inheritance. It will make available the modifier
+ * `onlyOwner`, which can be applied to your functions to restrict their use to
+ * the owner.
  */
-abstract contract Ownable is Context, ERC173 {
-    constructor(address initialOwner) ERC173(initialOwner) {}
+abstract contract Ownable is Context, IERC173, ERC165 {
+    address private _owner;
+
+    /// @dev The caller account is not authorized to perform an operation.
+    error OwnableUnauthorizedAccount(address account);
+
+    /// @dev The owner is not a valid owner account. (eg. `address(0)`)
+    error OwnableInvalidOwner(address owner);
 
     /**
-     * @dev Leaves the contract without owner. It will not be possible to call
-     * `onlyOwner` functions. Can only be called by the current owner.
-     *
-     * NOTE: Renouncing ownership will leave the contract without an owner,
-     * thereby disabling any functionality that is only available to the owner.
+     * @dev Throws if called by any account other than the owner.
      */
-    function renounceOwnership() public virtual onlyOwner {
-        _transferOwnership(address(0));
+    modifier onlyOwner() {
+        _checkOwner();
+        _;
+    }
+
+    /// @dev Initializes the contract setting the address provided by the deployer as the initial owner.
+    constructor(address initialOwner) {
+        if (initialOwner == address(0)) {
+            revert OwnableInvalidOwner(address(0));
+        }
+        _transferOwnership(initialOwner);
+    }
+
+    /// @inheritdoc ERC165
+    function supportsInterface(bytes4 interfaceId) public view virtual override(ERC165, IERC165) returns (bool) {
+        return interfaceId == type(IERC173).interfaceId || super.supportsInterface(interfaceId);
+    }
+
+    /// @inheritdoc IERC173
+    function owner() public view virtual returns (address) {
+        return _owner;
+    }
+
+    /// @inheritdoc IERC173
+    function transferOwnership(address newOwner) public virtual override onlyOwner {
+        _transferOwnership(newOwner);
+    }
+
+    /// @dev Throws if the sender is not the owner.
+    function _checkOwner() internal view virtual {
+        if (owner() != _msgSender()) {
+            revert OwnableUnauthorizedAccount(_msgSender());
+        }
     }
 
     /**
      * @dev Transfers ownership of the contract to a new account (`newOwner`).
-     * Can only be called by the current owner.
+     * Internal function without access restriction.
      */
-    function transferOwnership(address newOwner) public virtual override {
-        if (newOwner == address(0)) {
-            revert OwnableInvalidOwner(address(0));
-        }
-        super.transferOwnership(newOwner);
+    function _transferOwnership(address newOwner) internal virtual {
+        address oldOwner = _owner;
+        _owner = newOwner;
+        emit OwnershipTransferred(oldOwner, newOwner);
     }
 }