+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Breaking changes
 
 - Update minimum pragma to 0.8.24 in `Votes`, `VotesExtended`, `ERC20Votes`, `Strings`, `ERC1155URIStorage`, `MessageHashUtils`, `ERC721URIStorage`, `ERC721Votes`, `ERC721Wrapper`, `ERC721Burnable`, `ERC721Consecutive`, `ERC721Enumerable`, `ERC721Pausable`, `ERC721Royalty`, `ERC721Wrapper`, `EIP712`, and `ERC7739`. ([#5726](https://github.com/OpenZeppelin/openzeppelin-contracts/pull/5726))
+- `Ownable`'s `renounceOwnership` function was removed in favor of `transferOwnership(address(0))` in compliance with ERC-173. Users may use `OwnableRenounceable` instead as a drop-in replacement. Applies to all contracts that inherit from `Ownable`.
 
 ### Deprecation